+++ examples/buttons/main.go
@@ -15,6 +15,7 @@ import (
 	"strconv"
 	"time"
 
+	"github.com/gotd/log/logzap"
 	"go.uber.org/zap"
 
 	"github.com/gotd/botapi"
@@ -35,7 +36,7 @@ func menu() *botapi.InlineKeyboardMarkup {
 }
 
 func main() {
-	log, _ := zap.NewDevelopment()
+	log, _ := zap.NewProduction()
 	defer func() { _ = log.Sync() }()
 
 	appID, err := strconv.Atoi(os.Getenv("APP_ID"))
@@ -46,7 +47,7 @@ func main() {
 	bot, err := botapi.New(os.Getenv("BOT_TOKEN"), botapi.Options{
 		AppID:   appID,
 		AppHash: os.Getenv("APP_HASH"),
-		Logger:  log,
+		Logger:  logzap.New(log),
 	})
 	if err != nil {
 		log.Fatal("Create bot", zap.Error(err))