+++ examples/media/main.go
@@ -20,6 +20,7 @@ import (
 	"strconv"
 	"time"
 
+	"github.com/gotd/log/logzap"
 	"go.uber.org/zap"
 
 	"github.com/gotd/botapi"
@@ -29,7 +30,7 @@ import (
 const samplePhotoURL = "https://upload.wikimedia.org/wikipedia/commons/3/3a/Cat03.jpg"
 
 func main() {
-	log, _ := zap.NewDevelopment()
+	log, _ := zap.NewProduction()
 	defer func() { _ = log.Sync() }()
 
 	appID, err := strconv.Atoi(os.Getenv("APP_ID"))
@@ -40,7 +41,7 @@ func main() {
 	bot, err := botapi.New(os.Getenv("BOT_TOKEN"), botapi.Options{
 		AppID:   appID,
 		AppHash: os.Getenv("APP_HASH"),
-		Logger:  log,
+		Logger:  logzap.New(log),
 	})
 	if err != nil {
 		log.Fatal("Create bot", zap.Error(err))